Use ReadOnlyMemory<float> for all dense_vector types
Drops FloatVector / ByteVector wrappers, the JSON converters, ContextProvider,
JsonReader/WriterHelpers, the FloatVectorEncoding / ByteVectorEncoding enums,
EsqlVectorEncodingContext and the matching EsqlClientSettings knobs.

ES|QL uses float[] as the wire format for every dense_vector element_type
(float, byte, bit) in both directions (ingestion, params, responses), so a
single ReadOnlyMemory<float> covers all three. Knn and V_HAMMING collapse
to a single signature each. For byte / bit vectors users supply signed
whole-number floats in [-128, 127]; the server validates the range.

+++ tests/Elastic.Esql.Tests/Functions/Vector/VectorSimilarityTests.cs
@@ -37,7 +37,7 @@ public class VectorSimilarityTests : EsqlTestBase
 	{
 		var esql = CreateQuery<BookDocument>()
 			.From("books")
-			.Where(b => EsqlFunctions.VHamming(b.RgbVector, new byte[] { 1, 2, 3 }) < 10)
+			.Where(b => EsqlFunctions.VHamming(b.RgbVector, new float[] { 1, 2, 3 }) < 10)
 			.ToString();
 
 		_ = esql.Should().Contain("V_HAMMING(rgbVector, [1, 2, 3])");
